# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Forest Fire Prediction & Disaster Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List risk zones
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskZone"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a risk zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskZoneInput"
      responses:
        "201":
          description: Created zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskZone"
  /zones/{id}:
    patch:
      operationId: updateZone
      tags: [zones]
      summary: Update a risk zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskZoneUpdate"
      responses:
        "200":
          description: Updated zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskZone"
...
components:
  schemas:
...
    RiskZone:
      type: object
      required: [id, name, riskLevel, riskScore, latitude, longitude, radiusKm, forestType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        riskLevel:
          type: string
          enum: [low, moderate, high, extreme]
        riskScore:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        radiusKm:
          type: number
        forestType:
          type: string
          enum: [tropical, temperate, boreal, Mediterranean, savanna, mixed]
        populationDensity:
          type: ["string", "null"]
        lastAssessed:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RiskZoneInput:
      type: object
      required: [name, riskLevel, latitude, longitude, radiusKm, forestType]
      properties:
        name:
          type: string
          minLength: 1
        riskLevel:
          type: string
          enum: [low, moderate, high, extreme]
        riskScore:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        radiusKm:
          type: number
        forestType:
          type: string
          enum: [tropical, temperate, boreal, Mediterranean, savanna, mixed]
        populationDensity:
          type: string
        notes:
          type: string
    RiskZoneUpdate:
      type: object
      properties:
        riskLevel:
          type: string
          enum: [low, moderate, high, extreme]
        riskScore:
          type: number
        notes:
          type: string
        populationDensity:
          type: string
...