# lib/api-zod/src/generated/types/riskZone.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forest Fire Prediction & Disaster Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskZoneForestType } from './riskZoneForestType';
import type { RiskZoneRiskLevel } from './riskZoneRiskLevel';

export interface RiskZone {
  id: number;
  name: string;
  riskLevel: RiskZoneRiskLevel;
  riskScore: number;
  latitude: number;
  longitude: number;
  radiusKm: number;
  forestType: RiskZoneForestType;
  /** @nullable */
  populationDensity?: string | null;
  /** @nullable */
  lastAssessed?: Date | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}